requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.0
fastembed==0.4.2
jinja2==3.1.4
mcp>=1.0.0
psycopg2-binary>=2.9
neo4j>=5.14
redis>=5.0
